Added main functions to scripts and easy text display
removed the need for a separate security_cam.py and added main functions to MyCamera.py and
motion_sensor.py

Simplified text display with display_text function in MyCamera.py. When text is added with this
function, the function updates the offset needed for next  text display.

Motion sensor now correctly displays the movement intensity and signals when movement is found